01 -bitterbot-workspace/anthropic-build/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0
anthropic==0.21.3
python-multipart==0.0.6